requirements.txt
python-binance
google-genai
pandas
ta
python-dotenv
python-telegram-bot>=20.0
aiohttp
